This is version 3.0.1
Added new panel for FBX importing.
This panel is very convenient because the user
can see where is the current imported FBX file
coming from. The data of origin persists in the UI
and is also used to remember the working directories where
data is being exported to.

Overall it makes it very easy to work with individual characters and
keep track where the motions and actors are being exported to.

Signed-off-by: garrieta <garrieta@amazon.com>